Merged work-setup and work-setup-build
It now spawns work-setup-build as a separate frame. Renames the frame
names to Code and Shell to be able to easier tell them apart.

The old work-setup is now called work-setup-code. It also places the
Shell frame in the left screen and code on the right. It assumes a
screen is 1920 pixels wide.