Disabling auto close of *compilation* on success
It picks the first buffer in the buffer list, but doesn't check if any
of them are already active. So after a successful build you will end
up with two windows showing the same buffer.

Added that as a todo to fix it, in the meantime I've disabled that feature.