Disabled rtags
Just too slow to be useful. Takes a few seconds to jump to a
definition etc. It's faster for me to manually go to the definition
than it's to jump via rtags. Also launching rdm if it's not already
running from emacs wasn't a great idea. As that freezes emacs until
that's loaded, which takes 10-30s.