Switching over to company
Enabled company and tweaked it a lot to get it close to auto
complete. At work most of the backends are either too slow or just
fails half the time. So cannot have fancy auto complete parsed from a
compiler. Bonus is that ggtags doesn't messes up with the theme of the
popup.