{dd-insert-namespace} Root by WORKSPACE, added more options
It now looks for the WORKSPACE file to know where the start of the
project is. As swdevl wasn't that robust.

Also added that you can remove, replace and add namespaces instead of
just add.

Syntax for removing is prefix the entries with !, for example remove
the DD namespace add !DD.

Syntax for replacing is REGEX=REP. For example replace Geometry with
Geometry3D add Geometry=Geometry3D.