Cleaned up dd-insert-namespace{,-raw} and fixed dd-find-workspace
Realized that most functionality of the modify_namespace syntax can
describe the other hard coded stuff in dd-insert-namespace. So
refactored that into a new function dd-insert-namespace-raw. Then
changed dd-insert-namespace to just use the modify-namespace to
express the different rules and pass that into
dd-insert-namespace-raw.

Also discovered that directory-files can take a filter, so don't need
to use member to check if WORKSPACE exist or not. Can pass that
directly to the function.