Refactored dd-newfile into two files
Moved all the general functions for setting up the file into
newfile. And left the DD specific parts in dd-newfile. Changed the
layout a bit to avoid hardcoding the boilerplate etc. Instead it's
using three custom variables 'dd-boilerplate', 'dd-author-name' and
'dd-author-mail'. Which can be customize to use the right data.