newfile.el - Changed the namespace list to an alist
To cleaner handle the case for keyword, i.e. inline namespaces. Now
each elemet consist of a cons cell. Where the default of cdr is nil.

Refactored the nf-insert-include-guard to move the creation of the
include guard name to a separate function. It will also ignore
namespaces with associated keyword.