Remapped avy and created a new plattfot-mode for comint-mode buffers
As my plattfot-mode takes precedence over the comint major mode
searching back in history no longer works. Cleanest way I found was to
create a new minor mode that doesn't have that key binding. For that I
needed to make both buffer local, to be able change between them for
comint buffers. But to be able to enable my plattfot-mode by default
on all buffers I needed to make a global version of it. Therefore
there are three minor modes.