Fixed bug in ps-match when sorting based on time [config]
ps-match now return a timestamp when the process was started. Renamed
sort-ps-etime to sort-ps-time, as it now sorts based on timestamps.

Should have read the docs more carefully. An elisp time value that is
less than 1970 will be undefined (some systems might handle as low as
1901). As I was just converting the time difference to a time value I
was at year zero... Which didn't quite work with sort-ps-etime.

Switched to using lstart which returns <Name of day> <month> <day>
<time> <year>. That `date-to-time' can parse directly. No longer need
to have a custom function to parse the date string which is nice. Have
only tested this on cent6, note sure if lstart always return that
format. As the manual didn't specify what format it returns.