Switched to using flycheck for ledger [conf]
Tested flymake for linting ledger but did not like it. Mostly because
of how flymake behaves. I.e. There is no easy way of fixing the
modeline for flymake. One Flymake takes up a lot of space. But the
main issue I have is with the error and warning count. They use the
compilation faces for highlights and they stick out as a sore thumb in
my mode line with my current theme. Especially for inactive frames
where the whole mode line is gray except those two.

Other issue I have is that I don't get a quick info about the error
when the cursor is on the error. Instead I will either need to hover
my mouse pointer over it or bring up the diagnostic buffer. Granted
this is just a different workflow than flycheck. Would probably be
fine if I got use to it. The main beef I have with it is the lack of
customization of the mode line.