Fixed an issue with revert-all-buffers [conf]
It now skips buffers which file no longer exist. Got quite annoying
when reverting all buffers and you had multiple buffers which file
have been deleted. As everytime it encountered one it errored out and
stopped. To fix it you needed to visit that buffer and either close it
or write the buffer to the file. Both which might not be ideal.