Added better scrolling behavior for next/prior [conf]
I easily lost track of were I was on a page when using the default
binding for the next/prior. Which is why a rarely used them. But now I
got a kyria keyboard and I have bind page up/down to one of its rotary
encoders. And find it quite pleasant to use it to scroll.