Clean the org-agenda-list view [conf]
Added icons and removed the todo keywords, priority and tags. As
org-agenda-super already gives me that information.

Must be a better way of removing these things though. Right now I'm
just using the transformer in org-agenda-super to remove the from the
string.

Found settings `org-agenda-remove-tags` and
`org-agenda-todo-keyword-format` to remove the tags and todo keywords.
But they're global or I need to create a custom command which I don't
like.

Haven't found a setting for removing the priority.