Use arxiv-get-pdf-add-bibtex-entry to download papers [conf]
It seems that my code for updating the elfeed database with the extra
download url didn't stick.

Switching to using the arxiv code from `org-ref` instead. It's a bit
slower due to it needs to take some roundtrips to lookup the name of
the pdf. But this will add the bibtex info and the pdf name is a bit
more sane.