Workaround issue with ledger-mode and orderless smart case [conf]
See https://github.com/oantolin/orderless/issues/52 for full
explanation. But in short orderless smart case will be case sensitive
if it detects upper case in completion word. And ledger is sending the
full account when completing, which is triggering it. Workaround is
just to disable orderless smart case for ledger.