Shuffle around line and kill meow bindings and remove meow delete [conf]
Not using d and D for delete and delete backward, as backspace and
delete are in a good spot on my current keyboard.

Using x and s was a bit awkward so moving meow-line to s, and then
meow-kill to d. That way the key frequencies are more spread out on my
fingers.