Move repeat in meow to x [conf]
Pressing x to bring up M-x was nice, but it wasn't consistent. Where
as SPC m x works in all buffers except the mini-buffer.

Moving repeat to `x` as that is nicer to press than `'`. And will make
it more accessible on keyboards like ferris sweep as there `'` would
be on a layer or combo.